adds basic post routes and user migration
