+++ prisma/schema.prisma
@@ -10,18 +10,19 @@ datasource db {
   url      = env("DATABASE_URL")
 }
 
-// model User {
-//   id    Int     @id @default(autoincrement())
-//   email String  @unique
-//   name  String?
-//   posts Post[]
-// }
+model User {
+  id       String @id @default(uuid())
+  email    String @unique
+  name     String
+  password String
+  posts    Post[]
+}
 
 model Post {
-  id        Int     @id @default(autoincrement())
+  id        String  @id @default(uuid())
   title     String
   content   String?
   published Boolean @default(false)
-  // author    User    @relation(fields: [authorId], references: [id])
-  // authorId  Int
+  author    User    @relation(fields: [authorId], references: [id])
+  authorId  String
 }
